refactor: extract env toggle to separate component
